+++ Engine/Dude.h
@@ -9,13 +9,13 @@ public:
 	void Update(Keyboard& kbd);
 	void ClampToScreen();
 	void Draw(Graphics& gfx) const;
-	int GetX() const;
-	int GetY() const;
-	int GetWidth() const;
-	int GetHeight() const;
+	float GetX() const;
+	float GetY() const;
+	float GetWidth() const;
+	float GetHeight() const;
 private:
-	int x = 400;
-	int y = 300;
-	static constexpr int width = 20;
-	static constexpr int height = 20;
+	float x = 400;
+	float y = 300;
+	static constexpr float width = 20;
+	static constexpr float height = 20;
 };
