+++ Engine/Goal.cpp
@@ -1,6 +1,6 @@
 #include "Goal.h"
 
-Goal::Goal(int in_x, int in_y)
+Goal::Goal(float in_x, float in_y)
 {
 	x = in_x;
 	y = in_y;
@@ -8,10 +8,10 @@ Goal::Goal(int in_x, int in_y)
 
 void Goal::Draw(Graphics& gfx) const
 {
-	gfx.DrawRectDim(x, y, width, height, c);
+	gfx.DrawRectDim(int(x), int(y), int(width), int(height), c);
 }
 
-void Goal::Respawn(int xPos, int yPos)
+void Goal::Respawn(float xPos, float yPos)
 {
 	x = xPos;
 	y = yPos;
@@ -45,9 +45,9 @@ void Goal::UpdateColor()
 
 bool Goal::TestCollision(const Dude& dude) const
 {
-	const int right = x + width;
-	const int bottom = y + height;
-	const int dudeRight = dude.GetX() + dude.GetWidth();
-	const int dudeBottom = dude.GetY() + dude.GetHeight();
+	const float right = x + width;
+	const float bottom = y + height;
+	const float dudeRight = dude.GetX() + dude.GetWidth();
+	const float dudeBottom = dude.GetY() + dude.GetHeight();
 	return (x <= dudeRight && right >= dude.GetX() && y <= dudeBottom && bottom >= dude.GetY());
 }
