+++ Engine/Goal.h
@@ -5,16 +5,16 @@
 class Goal
 {
 public:
-	Goal(int in_x, int in_y);
+	Goal(float in_x, float in_y);
 	void Draw(Graphics& gfx) const;
-	void Respawn(int xPos, int yPos);
+	void Respawn(float xPos, float yPos);
 	void UpdateColor();
 	bool TestCollision(const Dude& dude) const;
 private:
-	int x;
-	int y;
-	static constexpr int width = 20;
-	static constexpr int height = 20;
+	float x;
+	float y;
+	static constexpr float width = 20.0f;
+	static constexpr float height = 20.0f;
 	Color c = { 127, 0, 0 };
 	bool isIncreasing = true;
 };
